refactor(vanity-url): Move VanityUrl model to Page namespace and update related tests
- Relocated VanityUrl class from Site to Page namespace for better organization.
- Updated PageAsset and tests to reflect the new VanityUrl structure.
- Enhanced README with additional information on VanityUrl properties and usage.
