+++ src/Model/Page/PageAsset.php
@@ -8,7 +8,6 @@ use Dotcms\PhpSdk\Model\Container\ContainerPage;
 use Dotcms\PhpSdk\Model\Content\Contentlet;
 use Dotcms\PhpSdk\Model\Layout\Layout;
 use Dotcms\PhpSdk\Model\Site\Site;
-use Dotcms\PhpSdk\Model\Site\VanityUrl;
 use Dotcms\PhpSdk\Model\View\ViewAs;
 
 class PageAsset
